# BS round up to BUCKET_SIZE to prevent capture graph when graph input not change