@BaihuiJin 2024-07-15
Actually If we could manage to only delete specific graph(decode stage), and keep the prefill graph, we could achieve a higher prefill throughput and a larger decode bs, but I couldn't find the right api : (