@kdamaszk 2024-07-15
@BaihuiJin you are right -- it would be much easier if we can manage particular graphs. However, I'm also not aware of such option